+++ idact/detail/jupyter_app/__init__.py
@@ -1,0 +1,2 @@
+"""This package contains implementation details of the quick Jupyter deployment
+ app, see :mod:`idact.notebook`."""